Dont display toolbar if no cache items

+++ administrator/components/com_cache/src/View/Cache/HtmlView.php
@@ -130,10 +130,13 @@ class HtmlView extends BaseHtmlView
 		// Get the toolbar object instance
 		$toolbar = Toolbar::getInstance('toolbar');
 
-		ToolbarHelper::custom('delete', 'delete', '', 'JTOOLBAR_DELETE', true);
-		ToolbarHelper::custom('deleteAll', 'remove', '', 'JTOOLBAR_DELETE_ALL', false);
-		$toolbar->appendButton('Confirm', 'COM_CACHE_RESOURCE_INTENSIVE_WARNING', 'delete', 'COM_CACHE_PURGE_EXPIRED', 'purge', false);
-		ToolbarHelper::divider();
+		if (count($this->data))
+		{
+			ToolbarHelper::custom('delete', 'delete', '', 'JTOOLBAR_DELETE', true);
+			ToolbarHelper::custom('deleteAll', 'remove', '', 'JTOOLBAR_DELETE_ALL', false);
+			$toolbar->appendButton('Confirm', 'COM_CACHE_RESOURCE_INTENSIVE_WARNING', 'delete', 'COM_CACHE_PURGE_EXPIRED', 'purge', false);
+			ToolbarHelper::divider();
+		}
 
 		if (Factory::getUser()->authorise('core.admin', 'com_cache'))
 		{
